extract a base menuview out of StavMenuView
